make format of the dates more granular
this is an experiment: when you first run the graphing script, there
are not many samples available. so everything looks like it's the same
time (e.g. "Year: 2015" for me. this is confusing. the above adds
months and days to the format, but also allows for mixing hours as
well.

this may not scale very well to large datasets. already with my ~24h
dataset, the processing is much much slower (7 seconds instead of ~0
seconds).